#443: Moved the PassiveEquipmentWeaponElementStatsIncreaseParser to the model folder
 -
