+++ src/app/ffbe/mappers/effects/passive-effect-parser.factory.ts
@@ -10,7 +10,7 @@ import {PassiveMpDecreaseForSongsParser} from './passives/passive-mp-decrease-fo
 import {PassiveEvasionMagicalParser} from './passives/passive-evasion-magical.parser';
 import {PassiveSkillMultipleActivationParser} from './passives/passive-skill-multiple-activation.parser';
 import {PassiveSkillTurnStartActivationParser} from './passives/passive-skill-turn-start-activation.parser';
-import {PassiveEquipmentWeaponElementStatsIncreaseParser} from './passives/passive-equipment-weapon-element-stats-increase.parser';
+import {PassiveEquipmentWeaponElementStatsIncreaseParser} from '../../model/effects/passives/stats/passive-equipment-weapon-element-stats-increase.parser';
 import {PassiveDeceivesDeathParser} from './passives/passive-deceives-death.parser';
 import {PassiveElementsAbsorbParser} from './passives/passive-elements-absorb.parser';
 import {PassiveStatsIncreaseUnarmedParser} from './passives/passive-stats-increase-unarmed.parser';
