+++ src/app/ffbe/model/effects/passives/stats/passive-equipment-weapon-element-stats-increase.parser.ts
@@ -1,5 +1,5 @@
-import {EffectParser} from '../effect-parser';
-import {Skill} from '../../../model/skill.model';
+import {EffectParser} from '../../../../mappers/effects/effect-parser';
+import {Skill} from '../../../skill.model';
 
 export class PassiveEquipmentWeaponElementStatsIncreaseParser extends EffectParser {
